Resolving conflicts

+++ hospital-equipment/src/main/java/com/ISAproject/hospitalequipment/Controller/EquipmentStockController.java
@@ -41,7 +41,7 @@ public class EquipmentStockController {
     //method for setting new amount instead of old
     @CrossOrigin(origins = "*")
     @PostMapping (value = "/update/{eqId}/{comId}")
-    public ResponseEntity<Void> updateAmount(@PathVariable("eqId") Long equipmentId,@PathVariable("comId")  Long comapnyId,@RequestParam("amount") int amount){
+    public ResponseEntity<Void> updateAmount(@PathVariable("eqId") Long equipmentId,@PathVariable("comId")  Long comapnyId,@RequestParam("amount") Long amount){
         this.equipmentStockService.updateAmount(equipmentId,comapnyId,amount);
 
         return new ResponseEntity<>( HttpStatus.OK);
